requirements.txt
beautifulsoup4
opencv-python
pandas
pytesseract
requests
rich
PyPDF2
python-docx
